refactor(purchase-orders): migrate components to feature directory and extract PurchaseOrderTable (Phase 6.1d)
- Move 'PurchaseOrderRow' from shared resources to 'features/purchase-orders/components'.
- Extract 'PurchaseOrderTable' component for consistent table orchestration.
- Update purchase orders list page to use new feature-specific components.
