+++ features/purchase-orders/components/index.ts
@@ -2,3 +2,5 @@ export * from './purchase-order-alert-dialogs';
 export * from './purchase-order-items-card';
 export * from './purchase-order-info-card';
 export * from './purchase-order-remarks-card';
+export { PurchaseOrderTable } from './purchase-order-table';
+export { PurchaseOrderRow } from './purchase-order-row';
